Fix FT_Memory
FT_Memory isa ?*struct_FT_MemoryRec_ and struct_FT_MemoryRec_ has 3 fields,
alloc, free and realloc which are pointer to functions each of which takes
a FT_Memory as its first parameter. This causes a compile "error:
'FT_Memory" depends on itself".

The solution is to chagne the first parameter to each of those *_Func's
from FT_Memory to ?*struct_FT_MemoryRec_.